pythonServe/requirements.txt
python-docx
beautifulsoup4
googletrans